+++ packages/patternfly-4/gatsby-theme-patternfly-org/components/commonComponents/commonComponents.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import { Link } from 'gatsby';
-import AutoLinkHeader from '../components/autoLinkHeader';
-import Example from '../components/example';
+import { AutoLinkHeader } from '../autoLinkHeader/autoLinkHeader';
+import { Example } from '../example/example';
 import './commonComponents.css';
 
 // These components replace how all elements in MDX are rendered
-// TODO: Copy all elements in https://github.com/patternfly/patternfly-next/blob/master/src/patternfly/components/Content/content.scss
 export const commonComponents = {
   inlineCode: Example,
   pre: React.Fragment,