+++ packages/patternfly-4/gatsby-theme-patternfly-org/layouts/index.js
@@ -1,0 +1,2 @@
+export * from './sideNavLayout/sideNavLayout';
+export * from './trainingLayout/trainingLayout';