+++ packages/patternfly-4/gatsby-theme-patternfly-org/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.9](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.8...gatsby-theme-patternfly-org@1.2.9) (2019-12-06)
+
+**Note:** Version bump only for package gatsby-theme-patternfly-org
+
+
+
+
+
+## [1.2.8](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.7...gatsby-theme-patternfly-org@1.2.8) (2019-12-06)
+
+**Note:** Version bump only for package gatsby-theme-patternfly-org
+
+
+
+
+
 ## [1.2.7](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.6...gatsby-theme-patternfly-org@1.2.7) (2019-12-04)
 
 **Note:** Version bump only for package gatsby-theme-patternfly-org